+++ topup/handler_test.go
@@ -75,8 +75,8 @@ func (suite *HandlerTestSuite) TestHandleMsgTopup() {
 			hmTypes.BytesToHeimdallAddress(addr.Bytes()),
 			fee,
 			txHash,
-			uint64(logIndex),
-			uint64(blockNumber),
+			logIndex,
+			blockNumber,
 		)
 
 		// handler
@@ -90,7 +90,7 @@ func (suite *HandlerTestSuite) TestHandleMsgTopup() {
 			hmTypes.BytesToHeimdallAddress(addr.Bytes()),
 			fee,
 			txHash,
-			uint64(logIndex),
+			logIndex,
 			uint64(blockNumber),
 		)
